fixed spoon version works good

+++ src/main/java/spoon/IncrementalLauncher.java
@@ -1,9 +1,9 @@
-/*
+/**
  * SPDX-License-Identifier: (MIT OR CECILL-C)
  *
- * Copyright (C) 2006-2023 INRIA and contributors
+ * Copyright (C) 2006-2019 INRIA and contributors
  *
- * Spoon is available either under the terms of the MIT License (see LICENSE-MIT.txt) or the Cecill-C License (see LICENSE-CECILL-C.txt). You as the user are entitled to choose the terms under which to adopt Spoon.
+ * Spoon is available either under the terms of the MIT License (see LICENSE-MIT.txt) of the Cecill-C License (see LICENSE-CECILL-C.txt). You as the user are entitled to choose the terms under which to adopt Spoon.
  */
 package spoon;
 
@@ -94,8 +94,8 @@ public class IncrementalLauncher extends Launcher {
 	}
 
 	private static void saveFactory(Factory factory, File file) {
-		try (FileOutputStream fileStream = new FileOutputStream(file)) {
-			new SerializationModelStreamer().save(factory, fileStream);
+		try {
+			new SerializationModelStreamer().save(factory, new FileOutputStream(file));
 		} catch (IOException e) {
 			throw new SpoonException("unable to save factory");
 		}
@@ -223,7 +223,7 @@ public class IncrementalLauncher extends Launcher {
 
 			Collection<CtPackage> oldPackages = oldFactory.Package().getAll();
 			for (CtPackage pkg : oldPackages) {
-				if (pkg.isEmpty() && !pkg.isUnnamedPackage()) {
+				if (pkg.getTypes().isEmpty() && pkg.getPackages().isEmpty() && !pkg.isUnnamedPackage()) {
 					pkg.delete();
 				}
 			}
